requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2024.12.14
click==8.1.8
colorama==0.4.6
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
drf-yasg==1.21.8
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.14
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uritemplate==4.1.1
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.1
